noticed small bug when refreshing on search page that cleared cuisine drop down, moved scope of fetching for cuisines
